rev: move forcing engine impl out of data provider source
